mergestyles file structure
