+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenuItem.deprecated.test.tsx
@@ -2,7 +2,7 @@ jest.mock('../../utilities/contextualMenu');
 
 import * as React from 'react';
 import { shallow, ShallowWrapper } from 'enzyme';
-import { ContextualMenuItem } from './ContextualMenuItem';
+import { ContextualMenuItemBase } from './ContextualMenuItem.base';
 import { IContextualMenuItemProps } from './ContextualMenuItem.types';
 import { IContextualMenuItem } from './ContextualMenu.types';
 import { IMenuItemClassNames } from './ContextualMenu.classNames';
@@ -21,7 +21,7 @@ describe('ContextMenuItemChildren', () => {
       onCheckmarkClick = jest.fn();
 
       wrapper = shallow(
-        <ContextualMenuItem
+        <ContextualMenuItemBase
           item={menuItem}
           classNames={menuClassNames}
           index={1}
@@ -58,7 +58,9 @@ describe('ContextMenuItemChildren', () => {
         menuItem = { key: '123', iconProps: { iconName: 'itemIcon' }, name: 'menuItem' };
         menuClassNames = getMenuItemClassNames();
 
-        wrapper = shallow(<ContextualMenuItem item={menuItem} classNames={menuClassNames} index={1} hasIcons={true} />);
+        wrapper = shallow(
+          <ContextualMenuItemBase item={menuItem} classNames={menuClassNames} index={1} hasIcons={true} />
+        );
       });
 
       it('renders the icon', () => {
@@ -75,7 +77,9 @@ describe('ContextMenuItemChildren', () => {
         menuItem = { key: '123', iconProps: {} };
         menuClassNames = getMenuItemClassNames();
 
-        wrapper = shallow(<ContextualMenuItem item={menuItem} classNames={menuClassNames} index={1} hasIcons={true} />);
+        wrapper = shallow(
+          <ContextualMenuItemBase item={menuItem} classNames={menuClassNames} index={1} hasIcons={true} />
+        );
       });
 
       it('renders the icon with iconName', () => {
@@ -94,7 +98,9 @@ describe('ContextMenuItemChildren', () => {
       menuItem = { key: '123', iconProps: {}, submenuIconProps: {} };
       menuClassNames = getMenuItemClassNames();
 
-      wrapper = shallow(<ContextualMenuItem item={menuItem} classNames={menuClassNames} index={1} hasIcons={true} />);
+      wrapper = shallow(
+        <ContextualMenuItemBase item={menuItem} classNames={menuClassNames} index={1} hasIcons={true} />
+      );
     });
 
     it('renders the menu icon', () => {
