+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenuItem.styles.ts
@@ -1,0 +1,26 @@
+import { IContextualMenuItemStyleProps, IContextualMenuItemStyles } from './ContextualMenuItem.types';
+// import {
+//   IStyle,
+//   ITheme,
+// } from '../../Styling';
+
+export const getStyles = (props: IContextualMenuItemStyleProps): IContextualMenuItemStyles => {
+  const {
+    className
+    // theme,
+  } = props;
+
+  // const { palette, semanticColors } = theme;
+
+  return {
+    root: [
+      'ms-ContextualMenuItem',
+      {
+        // Insert css properties
+      },
+      className
+    ]
+
+    // Insert className styles
+  };
+};
