+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenuItem.tsx
@@ -1,98 +1,17 @@
-import * as React from 'react';
-import { hasSubmenu, getIsChecked } from '../../utilities/contextualMenu/index';
-import { BaseComponent, getRTL } from '../../Utilities';
-import { Icon } from '../../Icon';
-import { IContextualMenuItemProps } from './ContextualMenuItem.types';
+import { styled } from '../../Utilities';
+import {
+  IContextualMenuItemProps,
+  IContextualMenuItemStyleProps,
+  IContextualMenuItemStyles
+} from './ContextualMenuItem.types';
+import { ContextualMenuItemBase } from './ContextualMenuItem.base';
+import { getStyles } from './ContextualMenuItem.styles';
 
-const renderItemIcon = (props: IContextualMenuItemProps) => {
-  const { item, hasIcons, classNames } = props;
-
-  const { iconProps } = item;
-
-  if (!hasIcons) {
-    return null;
-  }
-
-  if (item.onRenderIcon) {
-    return item.onRenderIcon(props);
-  }
-
-  return <Icon {...iconProps} className={classNames.icon} />;
-};
-
-const renderCheckMarkIcon = ({ onCheckmarkClick, item, classNames }: IContextualMenuItemProps) => {
-  const isItemChecked = getIsChecked(item);
-  if (onCheckmarkClick) {
-    const onClick = (e: React.MouseEvent<HTMLElement>) => onCheckmarkClick(item, e);
-
-    return <Icon iconName={isItemChecked ? 'CheckMark' : ''} className={classNames.checkmarkIcon} onClick={onClick} />;
-  }
-  return null;
-};
-
-const renderItemName = ({ item, classNames }: IContextualMenuItemProps) => {
-  if (item.text || item.name) {
-    return <span className={classNames.label}>{item.text || item.name}</span>;
-  }
-  return null;
-};
-
-const renderSecondaryText = ({ item, classNames }: IContextualMenuItemProps) => {
-  if (item.secondaryText) {
-    return <span className={classNames.secondaryText}>{item.secondaryText}</span>;
-  }
-  return null;
-};
-
-const renderSubMenuIcon = ({ item, classNames }: IContextualMenuItemProps) => {
-  if (hasSubmenu(item)) {
-    return (
-      <Icon
-        iconName={getRTL() ? 'ChevronLeft' : 'ChevronRight'}
-        {...item.submenuIconProps}
-        className={classNames.subMenuIcon}
-      />
-    );
-  }
-  return null;
-};
-
-export class ContextualMenuItem extends BaseComponent<IContextualMenuItemProps, {}> {
-  public render() {
-    const { item, classNames } = this.props;
-
-    return (
-      <div className={item.split ? classNames.linkContentMenu : classNames.linkContent}>
-        {renderCheckMarkIcon(this.props)}
-        {renderItemIcon(this.props)}
-        {renderItemName(this.props)}
-        {renderSecondaryText(this.props)}
-        {renderSubMenuIcon(this.props)}
-      </div>
-    );
-  }
-
-  public openSubMenu = (): void => {
-    const { item, openSubMenu, getSubmenuTarget } = this.props;
-    if (getSubmenuTarget) {
-      const submenuTarget = getSubmenuTarget();
-      if (hasSubmenu(item) && openSubMenu && submenuTarget) {
-        openSubMenu(item, submenuTarget);
-      }
-    }
-  };
-
-  public dismissSubMenu = (): void => {
-    const { item, dismissSubMenu } = this.props;
-    if (hasSubmenu(item) && dismissSubMenu) {
-      dismissSubMenu();
-    }
-  };
-
-  public dismissMenu = (dismissAll?: boolean): void => {
-    const { dismissMenu } = this.props;
-    if (dismissMenu) {
-      dismissMenu(undefined /* ev */, dismissAll);
-    }
-  };
-}
+/**
+ * ContextualMenuItem description
+ */
+export const ContextualMenuItem = styled<
+  IContextualMenuItemProps,
+  IContextualMenuItemStyleProps,
+  IContextualMenuItemStyles
+>(ContextualMenuItemBase, getStyles, undefined, { scope: 'ContextualMenuItem' });
